requirements.txt
azure-ai-vision
python-decouple

